# ktor-batterypack-example/src/main/resources/static/schema/api.yaml
openapi: 3.0.0
info:
  version: 0.0.1
  title: Florin API
  description: API definition
servers:
  - url: http://localhost:8080
paths:
...
  /api/plant-identification:
    post:
      description: Identify a plant from an image and provide care instructions
      operationId: IdentifyPlant
      tags:
        - plants
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
              properties:
                image1:
                  type: string
                  format: binary
                  description: Primary image of the plant to identify.
                image2:
                  type: string
                  format: binary
                  description: Additional image of the plant.
                image3:
                  type: string
                  format: binary
                  description: Additional image of the plant.
      responses:
        '200':
          description: Successful plant identification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plant'
        '422':
          description: Unable to identify plant from the provided image
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
  /api/plants:
    get:
      description: List plants for a temp identity
      operationId: ListPlants
      parameters:
        - name: params
          in: query
          required: false
          style: deepObject
          schema:
            type: object
            required:
              - paging
            properties:
              status:
                $ref: '#/components/schemas/PlantStatus'
              statusIn:
                type: array
                minItems: 1
                items:
                  $ref: '#/components/schemas/PlantStatus'
              id:
                type: string
                format: uuid
              createdAtLte:
                type: string
                format: date
              createdAtGte:
                type: string
                format: date-time
              paging:
                type: object
                required:
                  - limit
                  - offset
                properties:
                  limit:
                    type: integer
                    format: int64
                    maximum: 50
                  offset:
                    type: integer
                    format: int64
                    minimum: 0
      tags:
        - plants
      responses:
        '200':
          description: List of plants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantListing'
    post:
      description: Claim a plant by creating it from an identification and questionnaire.
      operationId: ClaimPlant
      tags:
        - plants
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlantCreate'
      responses:
        '200':
          description: Plant successfully claimed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plant'
        '404':
          description: Plant identification or temp identity not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Validation error.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PlantCondition:
      type: string
      description: Observable condition of the plant.
      enum:
        - HEALTHY
        - DECLINING
        - DISEASED
        - PEST_INFESTED
        - DEAD
      example: "HEALTHY"
    PlantConditionCause:
      type: string
      description: Identified or suspected cause of the condition.
      enum:
        - OVERWATERING
        - UNDERWATERING
        - SUNBURN
        - INSUFFICIENT_LIGHT
        - NUTRIENT_DEFICIENCY
        - ROOT_ROT
        - FUNGAL_INFECTION
        - BACTERIAL_INFECTION
        - PEST_DAMAGE
      example: "OVERWATERING"
    SoilType:
      type: string
      description: Type of commercially available soil mix.
      enum:
        - STANDARD
        - AROID
        - CACTUS
        - ORCHID
        - ACIDIC
        - AQUATIC
      example: "STANDARD"
    LightingType:
      type: string
      description: Ideal lighting placement.
      enum:
        - DIRECT
        - BRIGHT_INDIRECT
        - MEDIUM_INDIRECT
        - LOW
      example: "BRIGHT_INDIRECT"
    ToxicityLevel:
      type: string
      description: Overall toxicity level.
      enum:
        - NONE
        - LOW
        - MODERATE
        - HIGH
      example: "NONE"
    PlantDifficulty:
      type: string
      description: How difficult the plant is to keep alive.
      enum:
        - BEGINNER
        - EASY
        - MODERATE
        - EXPERT
      example: "EASY"
...
    ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type.
          example: "https://example.com/probs/out-of-credit"
        title:
          type: string
          description: A short, human-readable summary of the problem type.
          example: "You do not have enough credit."
        status:
          type: integer
          format: int32
          description: The HTTP status code generated by the origin server for this occurrence of the problem.
          example: 403
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
          example: "Your current balance is 30, but that costs 50."
        instance:
          type: string
          format: uri
          description: A URI reference that identifies the specific occurrence of the problem.
          example: "/account/12345/transactions/abc"
    PlantImageInfo:
      type: object
      required:
        - id
        - url
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the image.
        url:
          type: string
          description: URL to fetch the image.
    PlantIdentificationItem:
      type: object
      required:
        - species
        - speciesProbability
      properties:
        species:
          type: string
          description: Scientific name of the species.
        mostLikelyCommonName:
          type: string
          description: The most likely common name for the plant, with cultivar if applicable.
        cultivar:
          type: string
          nullable: true
          description: Cultivar or variety name, if applicable.
        commonNames:
          type: array
          items:
            type: string
          description: List of common names for the plant.
        speciesProbability:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: AI confidence for the species (0.0 to 1.0).
        cultivarProbability:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: AI confidence for the cultivar (0.0 to 1.0).
    PlantHealthStatus:
      type: object
      required:
        - condition
      properties:
        condition:
          $ref: '#/components/schemas/PlantCondition'
        cause:
          $ref: '#/components/schemas/PlantConditionCause'
        details:
          type: string
          description: Detailed explanation of the health status.
    PlantCareInstructions:
      type: object
      required:
        - watering
        - soil
        - repotting
        - lighting
        - fertilizing
      properties:
        watering:
          $ref: '#/components/schemas/PlantMonthlySchedule'
        soil:
          $ref: '#/components/schemas/PlantSoil'
        repotting:
          $ref: '#/components/schemas/PlantRepotting'
        lighting:
          $ref: '#/components/schemas/PlantLighting'
        fertilizing:
          $ref: '#/components/schemas/PlantMonthlySchedule'
    PlantMonthlySchedule:
      type: object
      required:
        - jan
        - feb
        - mar
        - apr
        - may
        - jun
        - jul
        - aug
        - sep
        - oct
        - nov
        - dec
      properties:
        jan:
          type: integer
          description: Days between actions in January.
        feb:
          type: integer
          description: Days between actions in February.
        mar:
          type: integer
          description: Days between actions in March.
        apr:
          type: integer
          description: Days between actions in April.
        may:
          type: integer
          description: Days between actions in May.
        jun:
          type: integer
          description: Days between actions in June.
        jul:
          type: integer
          description: Days between actions in July.
        aug:
          type: integer
          description: Days between actions in August.
        sep:
          type: integer
          description: Days between actions in September.
        oct:
          type: integer
          description: Days between actions in October.
        nov:
          type: integer
          description: Days between actions in November.
        dec:
          type: integer
          description: Days between actions in December.
        details:
          type: string
          description: Additional details.
    PlantSoil:
      type: object
      required:
        - type
      properties:
        type:
          $ref: '#/components/schemas/SoilType'
        details:
          type: string
          description: Additional details about the soil mix.
    PlantRepotting:
      type: object
      required:
        - intervalMonths
        - recommendedMonths
      properties:
        intervalMonths:
          type: integer
          description: Recommended number of months between repottings.
        recommendedMonths:
          type: array
          items:
            type: integer
            minimum: 1
            maximum: 12
          description: Months of the year (1-12) when repotting is safe.
        details:
          type: string
          description: Additional details about repotting.
    PlantLighting:
      type: object
      required:
        - type
      properties:
        type:
          $ref: '#/components/schemas/LightingType'
        details:
          type: string
          description: Additional details about lighting needs.
    PlantToxicity:
      type: object
      required:
        - level
      properties:
        level:
          $ref: '#/components/schemas/ToxicityLevel'
        details:
          type: string
          description: Additional details about toxic parts, symptoms, or precautions.
...
    LastWateredOption:
      type: string
      description: When the plant was last watered.
      enum:
        - TODAY
        - ONE_TO_TWO_DAYS
        - THREE_TO_FIVE_DAYS
        - MORE_THAN_WEEK
        - UNKNOWN
    SoilMoistureOption:
      type: string
      description: Current soil moisture level.
      enum:
        - WET
        - MOIST
        - SLIGHTLY_DRY
        - DRY
        - VERY_DRY
    YesNoOption:
      type: string
      description: Yes or no answer.
      enum:
        - "YES"
        - "NO"
    LightLocationOption:
      type: string
      description: Light exposure location for the plant.
      enum:
        - FAR_FROM_WINDOW
        - BRIGHT_ROOM
        - NEXT_TO_WINDOW
        - DIRECT_SUN
    PlantCreate:
      type: object
      required:
        - plantId
        - displayName
        - lastWatered
        - soilMoisture
        - repottedSinceBuying
        - lightLocation
        - nearHeatOrAc
      properties:
        plantId:
          type: string
          format: uuid
          description: ID of the plant identification to claim.
        displayName:
          type: string
          description: User-chosen display name for the plant.
        lastWatered:
          $ref: '#/components/schemas/LastWateredOption'
        soilMoisture:
          $ref: '#/components/schemas/SoilMoistureOption'
        repottedSinceBuying:
          $ref: '#/components/schemas/YesNoOption'
        lightLocation:
          $ref: '#/components/schemas/LightLocationOption'
        nearHeatOrAc:
          $ref: '#/components/schemas/YesNoOption'
    Plant:
      type: object
      required:
        - id
        - createdAt
        - status
        - identifications
        - healthStatus
        - careInstructions
        - difficulty
        - toxicity
        - images
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/PlantStatus'
        identifications:
          type: array
          description: List of potential plant identifications sorted by probability.
          items:
            $ref: '#/components/schemas/PlantIdentificationItem'
        displayName:
          type: string
          description: User-friendly name for the plant, based on the most likely identification.
        species:
          type: string
          description: Scientific name of the species, based on the most likely identification.
        cultivar:
          type: string
          description: Cultivar or variety name, if applicable, based on the most likely identification.
        healthStatus:
          $ref: '#/components/schemas/PlantHealthStatus'
        careInstructions:
          $ref: '#/components/schemas/PlantCareInstructions'
        difficulty:
          $ref: '#/components/schemas/PlantDifficulty'
        toxicity:
          $ref: '#/components/schemas/PlantToxicity'
        images:
          type: array
          items:
            $ref: '#/components/schemas/PlantImageInfo'
          description: List of images associated with the plant.
        tempIdentityId:
          type: string
          format: uuid
          description: Temporary identity ID for users without an account.
        lastWatered:
          $ref: '#/components/schemas/LastWateredOption'
        soilMoisture:
          $ref: '#/components/schemas/SoilMoistureOption'
        repottedSinceBuying:
          $ref: '#/components/schemas/YesNoOption'
        lightLocation:
          $ref: '#/components/schemas/LightLocationOption'
        nearHeatOrAc:
          $ref: '#/components/schemas/YesNoOption'
        lastWateringDate:
          type: string
          format: date
          description: Date of the last watering, calculated from lastWatered option at claim time.
        lastFertilizationDate:
          type: string
          format: date
          description: Date of the last fertilization, null if unknown.
        lastRepottingDate:
          type: string
          format: date
          description: Date of the last repotting, set to createdAt if repottedSinceBuying is TRUE.
        lastStatusUpdateDate:
          type: string
          format: date
          description: Date of the last status update, null if no status update yet.
    PlantStatus:
      type: string
      enum:
        - IDENTIFICATION
        - CLAIMED_PLANT
      description: Status of the plant record.
    PlantListing:
      type: object
      required:
        - total
        - items
      properties:
        total:
          type: integer
          format: int64
          description: Total number of plants.
          example: 5
        items:
          type: array
          items:
            $ref: '#/components/schemas/Plant'
          description: List of plants.
...